Massive refactor removing the memory caches (#113)
* Huge commit, remove the memory caches

Texture, skeleton, animation, shader, mesh, model and vertex layout isn't using the resource system for code allocated resources, but instead use the standard CreateX functions to create what is a non-resource type version of that resource. What that means is that the resource can't go through the resource system for discarding, but has to be explicitly deleted.
As a result, the resource system has gotten a refactor which means a few things.
1. There is no PrepareLoad, instead there is a LoadImmediate and LoadDeferred. LoadDeferred either loads the resource a bit later, or loads it on a thread, depending on what the loader supports.
2. The resource loader returns a ResourceUnknownId, which is a 24_8 bit resource id, unlike ResourceId which contains the loader and reference index. Internally, the resource id is updated with the one from the loader and is either returned directly if immediate load is performed, or in the callbacks if the loading is deferred. The result of the load is based on the ID returned, if invalid, the load is considered failed, if valid, it's considered a success.
3. 2. Might change, as we might want to be able to have the loader tell the resource system that the load neither failed nor succeeded, but that there is some limit on the resource system. For example, let's say we don't have enough upload buffer memory left this frame, then we might want to delay the load to next frame, and that should definitely be supported by the loaders.

All of this also mean that where previously we could assign TextureId from CreateResource, we now have to assign ResourceId, as TextureId is 24_8 but ResourceId is 24_8_24_8. The typed resources can be assigned from ResourceId, but not the other way around, so to discard a resource you'd need the original ResourceId. This is because when calling CreateTexture, you can't really return a full ResourceId since it's not associated with any loader. This API is similar to the ManagedResource in the old Nebula, except this time both the stream and memory loaders both reference the same source allocator.

The 'caches' have been renamed to loader to better match what they do, even if they also handle ResourceIds in the sense that they call the associated Destroy function when the usage count drops to 0.

* Fix unsigned ints.

* Fix material LODs getting streamed in.