Refactor frame script to be JSON compliant
Added a framescript schema for validation and assistance in writing framescript operations.
Changed ImageAspect to ImageBits which is a better name.
Fixed synchronization barrier for Bloom.
Changed the default render texture layout from shader read only to depth/color attachment read, this is to prevent unnecessary decompressions triggered by transitioning render textures.
Removed the use of the Auto ImageBits enum, forcing it to only be used for setup.
Added stencil format check for pixel format, along with a way to convert the format to image bits.