Add custom model (not from resource system)
We needed an API to just render a mesh with visibility and everything enabled. This was previously impossible because the ModelContext entirely relied on model resource type structures to work.

One of the major changes here is that there is no longer any primitive group or mesh bind callbacks. Instead, the ModelContext accepts a MeshId and a primitive group index instead. The actual primitive group is extracted by a job and put in the a visibility draw command later.

This made it possible to create a new Setup for ModelContext that simply accepts a material, mesh, transform and bounding box.

Furthermore, this simplifies the model node code as well as they don't have to implement custom callbacks to bind the mesh.