Area lights (#148)
* Fixed json reader not handling empty nodes.

* Small optimization to bbox clip status to test against orthogonal projection

* Add support for long list of appends with array and arraystack.

* Remove dependency on node in renderables.

Motivation here is that we want to avoid having to pass the node pointer everywhere.

However, we still need it for the skin fragments, so I have to figure out a way to fix that...

* Removed bounding box from primitive group.

* Visibility takes use of the ortho flag for bbox clip status.

* Added support for triangle fan.

* Simplified mesh container.

We don't need to have all attributes in separate arrays since they are always accessed all at the same time.

* Add custom model (not from resource system)

We needed an API to just render a mesh with visibility and everything enabled. This was previously impossible because the ModelContext entirely relied on model resource type structures to work.

One of the major changes here is that there is no longer any primitive group or mesh bind callbacks. Instead, the ModelContext accepts a MeshId and a primitive group index instead. The actual primitive group is extracted by a job and put in the a visibility draw command later.

This made it possible to create a new Setup for ModelContext that simply accepts a material, mesh, transform and bounding box.

Furthermore, this simplifies the model node code as well as they don't have to implement custom callbacks to bind the mesh.

* Small optimization in how we fetch shader configs.

* Added some geometry setup helpers to create meshes.

Added a disk and a rectangle mesh as two global meshes.

* Added support for LTC area lights!

Currently supports disk and rect lights.

Tube lights will be added later.

Also refactored a bunch of things in the shaders.

There is no PostEffectSampler or PostEffectUpscaleSampler. These have been renamed to PointSampler and LinearSampler respectively. Makes a lot more sense.

lights_clustered.fxh have been renamed to lighting_functions.fxh as it only contains light shading code, nothing to do with clustering really.

Added an emissive shader for area lights mesh proxy, as well as an area light material and pass in the vkdefault.json frame script.

* Fix build.

* Add xyz fetch to point.

* Change transform44 to use points for position and pivots

* Add API for getting a single primitive group from mesh

* Many area light changes

Added support for tube lights.
Changed visibility proxy to use a box instead of a frustum for all lights.
Changed LightContext to use position, rotation and scale as three separate parameters instead of using transforms directly.
Renamed global light to directional light in case we want to add more of them.

* Move bounding box clip bits down so they start at 0.

* Make sure mesh locks and uses safe getters

* Make sure to move all components of an entity.

* Revert to using the lights function that does scalarization

Also clamp light values to 60k to avoid overflowing 10-11-10 range.

* Use rectangle, disk and tube lights in clustered lighting example

* Fix compile

* Remove clamping on roughness.