New API for creating pipelines directly on the graphics device
Pipelines should be first class citizens and we should really move away from using the SetProgram, SetVertexLayout, etc approach to a more pipeline oriented one. This PR allows for creations of pipelines from outside the render loop, as well as setting them on the command buffer, which is a first necessary step to avoid runtime pipeline creation.