Fix various validation errors
Samplers are now hashed, preventing us from recreating the same sampler multiple times.
Downsampling context avoids global binds when running the shaders, as that throws a validation error as well as it is redundant.
The shader also properly counts the amount of resources being bound to all shader stages when no visibility is given.
Added a hash combine method so we can do more complicated hashes in the future.