Change how render pass attachment store/load/discard bits are defined.
Add discard bit
Change frame script schema to accept a string list (concatenated with |) to define if the attachment should load, store and/or discard. Clear is set by the clear value.
Add support for Vulkan 1.3 render pass store op which doesn't allow the GPU to discard the contents after the pass ends.