Created a few barrier helpers to make sure barriers can be reversed easily
Also fixed what I suspect to be a problem with the transfer. It's now handled by an RAII lock instead, which guarantees both the transfer and setup command buffers are locked at the same time. Previously, it would be possible that transfers was done in one frame and graphics in another, which is invalid.