Start working on a different material system.
We want all the materials on the GPU, so we need a new material system to do so.
Material 'signatures' are handled CPU side instead of being purely content driven. Requires a bit of extra work if one wants a new material, but gives full control over materials passed to the GPU.
Resource loading is seeing a small but significant change in transfer operations not waiting or being waited for. This is to be coupled with better sparse handling, enabling us to have growing GPU buffers work in tandem with transfer operations.
Sparse buffers have been implemented.