More compile time materials
Make material compiler build all the varibles and per-pass bindings as statics, such that we can access them directly in the code. One example of using this is the LightContext when setting up an area light.

It also keeps all the data in static memory, which might be better cache-wise than saving them directly in dictionaries.