New material setup for raytracing (#208)
* Initial commit for DDGI

* Move raytracing shaders to their folders.

This changes move the ray tracing shaders out of syswork to their constituent code folders. There has been some changes in the CMake scripts for this to work, mainly with how shaders are located and output.

* Start working on a different material system.

We want all the materials on the GPU, so we need a new material system to do so.
Material 'signatures' are handled CPU side instead of being purely content driven. Requires a bit of extra work if one wants a new material, but gives full control over materials passed to the GPU.
Resource loading is seeing a small but significant change in transfer operations not waiting or being waited for. This is to be coupled with better sparse handling, enabling us to have growing GPU buffers work in tandem with transfer operations.
Sparse buffers have been implemented.

* Fix range allocator allocating an extra unnecessary byte.

* Add material template compiler

* Start working on new material templates.

* Material template generation only on non-virtual materials.

* Remove enum output from materials, include material templates in material loader.

* Add back enum, change namespace to MaterialTemplates

* Once more revamped shader compilation.

The shader compiler doesn't make any folder assumptions at all. The CMake script takes care of providing proper paths to the shader compiler, but explicitly groups shaders in work under render/res/shaders.
All engine shaders ends up in a folder named system_shaders. This is to avoid other projects compiling shaders with equal names overwriting engine shaders.
Shader folder hierarchy is traversed and each level of folders are output to both the .fbx and the headers in order to avoid file collisions.

* Continue material codegen.

* Remove prints

* Continue work on new materials.

Allow for multiple json files to be merged into one single source file at the end. This allows extensions of materials in an easy manner.
Add legacy material for the MaterialParams struct we have already in Nebula. Easier to transition using it.
Replace material load code with new compile time version.
Make generators create untabbed files.

* Use a 64 bit mask to indicate where uniform buffers are bound to the shader instead of having it defined entirely dynamic.

Makes it easier to find uniform binding slots. 64 spaces should be enough.

* Fix a bunch of small things related to setup for materials

Also fix shader server looking for shared.fxb, which has been moved to system_shaders/shared.fxb

* Start grouping resources using bit masks in shaders.

Instead of providing arrays and dictionaries, it should suffice to use a 64 bit mask to describe each binding per set. Obviously this limits the size of our sets, but 64 bindings should definitely suffice.

* Fix new shader paths

* Fix global constants initialization sequence

* Update material stuff

* Fix stringatom hash

* Fix texture upload issues related to immediate loading

* Fix issues related to material binding

TODO: When resizing arrays, we don't initialize all new values to 0, this causes some of the binding masks to get borked. Have to fix this.

* When immediately loading, we should set the loaded bits to the requested bits and skip the streaming step altogether.

This method is a bit brute force and can technically fail, so there should be some mechanism to delay mips that were not loaded if this happens, even for immediate resources.

Alternatively, immediate resources can allocate an intermediate buffer. That might actually be more stable.

* More material work

Also added a resize with default value constructors for Array

* Added message print for batches which are incorrectly defined.

This will spam quite a lot, so we should find some other way of finding batches which has no material associated with it instead.

* Use alloca for barrier allocation.

* Fix timeline poll

* Fix typo

* Fixes related to streaming

Allow for resources to be immediately loaded if immediate is used.
Add another submission type for handovers, so graphics don't have to wait for transfers in the same frame.

* Fix a bunch of material related issues

* Use precompiled material templates in frame system.

Batches not corresponding to any material template will be flagged as useless.

* More compile time materials

Make material compiler build all the varibles and per-pass bindings as statics, such that we can access them directly in the code. One example of using this is the LightContext when setting up an area light.

It also keeps all the data in static memory, which might be better cache-wise than saving them directly in dictionaries.

* Use coregraphics flag to check raytracing support in game layer

* Few BLAS fixes

* Put DeviceAddress in CoreGraphics namespace

* Add support for 16 bit integers if raytracing is used

* Kill the raytracing library approach

Instead, just setup the entire raytracing pipeline from the shader modules found in each shader program provided.

* Fix uint when uchar is sufficient

* Fix missed DeviceAddress namespace

* Various material related shader updates

* Flush material buffer on both compute and graphics queues (whichever comes first) before submission

* Give internal buffers in BufferWithStaging names

* Use 1 byte mask for raytracing registration

* Fix barycentrics in raytracing and work on local lights.

* Fix mesh loader potentially reuploading the same mesh twice

* Generalize blas creation API and start supporting terrain with RT

* Submit all pending changes to merge in to main.

* Move raytracing shaders to their folders.

This changes move the ray tracing shaders out of syswork to their constituent code folders. There has been some changes in the CMake scripts for this to work, mainly with how shaders are located and output.

* Fix bad merge

* Fix more bad merge

* Fix compile issues

* Fix alloca for posix.

* Stupid mistake fix