New material system refactor
Materials now define their shader interface which results in code generated for shaders to fetch material parameters from. This guarantees having only one constant buffer per material, together with a fixed binding, which makes material setup significantly easier.
This also allows all materials to alias the same constant buffer in the shader, which allows for the same material parameters to be used for other shaders sharing parameters, such as shadows.
Textures are also treated in the same manner, with having a fixed set of slots, and a maximum of 16 bound textures per material.

As of this change, slots 51 and up are reserved for material bindings, which should give plenty of space for other resources.