Working on a better resource load process
Resource loaders are now responsible for allocating their own command buffers for uploads, handovers and whatnot, and then adding them to a list for submission. This removes intermittent locks and should fix a bug where resource loaders don't appropriately wait for the right submission index to mark resources as finished.