Complete rewrite of async loader system
The new system produces outputs from each loader thread which are then fed back to the main thread for further processing. Any partially completed jobs will be reissued with the bits that are left. No loader state is allowed to be changed in the job, eliminating any race conditions. The new system also provides a significantly more robust way of knowing when a resource is done.

The texture and mesh loaders have been significantly changed to record their own transfer commands in the threads. The submission of these commands is handed over to the main thread, and the subsequent cleanup of said commands happens on the threads at a later stage. This way we don't run into multiple threads accessing the command buffer pools. When the deletion happens, the output from the loader is that the resource is finished, thereby ending the streaming sequence.

TODO: LOD updates only run when a resource is finished loading. This prevents the loader from having to deal with multiple entries of the same resource out of order. This could be improved by guarding against unnecessary stream loads already on the main thread, with the loader output feeding back which LOD was successfully loaded last. It's not necessary to do so, but it might improve performance somewhat because resource loaders don't need to run on jobs which will produce no new output.