Refactor how GPU uploads are handled.
Previously, the graphics device would free up all upload slots each frame. With the new way we handle resources, which is fully async, these memory ranges will have to be freed up and flushed when we are actually done with them.

This change has the graphics device return the allocation to the user, which should then be managed by them to avoid upload memory getting permanently stuck. The texture loader and mesh loader take on this responsibility by retaining the allocation ranges when an upload is performed, and frees them up when the command buffer needing them is finished executing and is deleted.