Reverted to using a basic freeids array in the IdAllocator.
Reason for this is that the IdGenerationPool is not meant for this purpose. The purpose of the IdGenerationPool is to generate Ids that is supposed to be used in a lookup table (an array) to get the actual index into the linear buffer we want to associate it with. thus, it doesn't fit the IdAllocator which is more like a SOA container.