Fix resource cross queue barriers.
If framescripts would apply a barrier between for example graphics and compute queues, the compute queue would take the source stage, which might be a graphics specific stage (pixel shader write, for example) and try to sync with the compute side. Some pipeline stages are unique to graphics queues, therefore such barriers would be invalid.

Now, the frame script demotes the source stage to one that is compatible for that queue.