Multiple editor improvements
Adding a window which shows the progress of batching assets on the fly.
Terrain saves it's height map and biome maps to disk.
Added support in our NIDL for serializing Util::Array.
Added method in String to remove a substring from the start, to get a relative folder path.
Added a 16 bit png saver routine for STB.
Streamed/Loaded textures are now automatically created with transfer source for copies.